# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is FILKOM Canteen API Documentation",
        "title": "FILKOM Canteen API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5700",
    "paths": {
...
        "/api/v1/shops": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Fetch All Shops From Database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops (Admin only)"
                ],
                "summary": "Fetch All Shops",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginlib.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Shop"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Register Shop to System",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops (Admin only)"
                ],
                "summary": "Register Shop",
                "parameters": [
                    {
                        "description": "Shop Register Payload",
                        "name": "ShopPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ShopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/shops/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Fetch Shop By ID From DB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops (Admin and Owner)"
                ],
                "summary": "Fetch Shop By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginlib.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Shop"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginlib.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Shop"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Update Existing Shop",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops (Admin and Owner)"
                ],
                "summary": "Update Shop",
                "parameters": [
                    {
                        "description": "Shop Register Payload",
                        "name": "ShopPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ShopRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Shop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginlib.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Shop"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Username already exists",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Delete Existing Shop",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops (Admin only)"
                ],
                "summary": "Delete Shop",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/shops/{id}/owners/{ownerId}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Add Owner to Shop",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops (Admin only)"
                ],
                "summary": "Add Owner to Shop",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Owner ID",
                        "name": "ownerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    },
                    "404": {
                        "description": "Shop or Owner not found",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Remove Owner from Shop",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops (Admin only)"
                ],
                "summary": "Remove Owner from Shop",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Owner ID",
                        "name": "ownerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginlib.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Shop": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "shop_description": {
                    "type": "string"
                },
                "shop_id": {
                    "type": "string"
                },
                "shop_name": {
                    "type": "string"
                },
                "shop_photo_link": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.ShopRequest": {
            "type": "object"
        },
        "ginlib.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for accessing all endpoints. Type: Key TOKEN",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        },
        "UserAuth": {
            "description": "API Key for accessing protected user and admin endpoints. Type: Bearer TOKEN",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}